refactor(core): own ShellRun store invariants once
The record shape, patch, and transition rules were private to the SQLite
store, so a second ShellRunStore would have had to copy the state machine
rather than only its storage medium. Move them to @maka/core next to the
existing ShellRun predicates and have the SQLite store consume them.
