+++ packages/core/src/index.ts
@@ -513,6 +513,12 @@ export {
   SHELL_RUN_ACTIVE_STATUSES,
   SHELL_RUN_STATUSES,
   SHELL_RUN_TERMINAL_STATUSES,
+  assertShellRunIdentifier,
+  assertShellRunPatch,
+  assertShellRunSessionId,
+  nextShellRunRecord,
+  normalizeShellRunRecord,
+  shellRunNotFoundError,
   isShellOutput,
   isActiveShellRunStatus,
   isShellRunId,
